simplify widget routes
