+++ docs/openapi/paths/index.yaml
@@ -48,8 +48,6 @@
   $ref: './files.yaml'
 '/grantee/{granteeId}':
   $ref: './grantee/grantee-id.yaml'
-'/widgets/dashboardOverview':
-  $ref: './widgets/dashboardOverview.yaml'
 '/widgets/overview':
   $ref: './widgets/overview.yaml'
 '/widgets/reasonList':
