+++ frontend/src/widgets/DashboardOverview.js
@@ -131,4 +131,4 @@ DashboardOverviewWidget.defaultProps = {
   ],
 };
 
-export default withWidgetData(DashboardOverviewWidget, 'dashboardOverview');
+export default withWidgetData(DashboardOverviewWidget, 'overview');
