+++ src/widgets/index.js
@@ -2,7 +2,6 @@ import example from './example';
 import overview from './overview';
 import totalHrsAndGranteeGraph from './totalHrsAndGranteeGraph';
 import reasonList from './reasonList';
-import dashboardOverview from './dashboardOverview';
 import topicFrequencyGraph from './topicFrequencyGraph';
 
 /*
@@ -13,6 +12,5 @@ export default {
   overview,
   totalHrsAndGranteeGraph,
   reasonList,
-  dashboardOverview,
   topicFrequencyGraph,
 };
